Fixed Bug with game.makeMove
makeMove wasn’t emptying the location that the piece came from. This
meant that even though it wasn’t being drawn, the board that a move
duplicated a piece, putting the copy in the end location and leaving
the original at the start location. Fixed by simply having makeMove set
board[start] = nil. Dad, can you check and make sure that this is the
proper way to fix it?